create enemy spawn manager
