+++ Assets/Scripts/Enemy/StraightMovementStrategy.cs
@@ -4,7 +4,7 @@ public class StraightMovementStrategy : IMovementStrategy
 {
     public Vector2 CalculateMovementDirection(Transform enemyTransform)
     {
-        // Move in the direction the enemy is facing
+        
         return enemyTransform.right;
     }
 }
